+++ tests/conftest.py
@@ -1,8 +1,14 @@
 from itertools import chain, product
 from pathlib import Path
-from typing import Callable, Iterator, List, Tuple
+import sys
+from typing import List, Tuple
 
 import pytest
+
+if sys.version_info >= (3, 9):  # pragma: no cover
+    from collections.abc import Callable, Iterator
+else:  # pragma: no cover
+    from typing import Callable, Iterator
 
 
 def pytest_addoption(parser: pytest.Parser) -> None: