+++ CHANGELOG.md
@@ -15,6 +15,7 @@ adheres to [Semantic Versioning](https://semver.org/).
 - Refactor a descriptor following PEP 487
 - Add tests for CPython 3.11 and PyPy 3.9
 - Use CPython 3.11 for misc. tests
+- Update Github actions dependencies
 
 ## [0.4.0] - 2022-01-09