+++ CHANGELOG.md
@@ -11,7 +11,7 @@ adheres to [Semantic Versioning](https://semver.org/).
 
 ### :house: Internal
 
-- Necessary code changes following dev dependency update: black, pytest
+- Necessary code changes following dev dependency update: black, pylint, pytest
 - Refactor a descriptor following PEP 487
 - Add tests for CPython 3.11 and PyPy 3.9
 - Use CPython 3.11 for misc. tests