+++ CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project
 adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.5.0] - 2023-02-27
 
-[unreleased]: https://github.com/rogdham/python-xz/compare/v0.4.0...HEAD
+[0.5.0]: https://github.com/rogdham/python-xz/releases/tag/v0.5.0
 
 ### :boom: Breaking changes